Move css to styles directory
